Make a forms folder for form components and extract interval chart into own component
