+++ client/src/components/forms/EditFlagForm.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useParams } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
-import { editFlag } from "../actions/flagActions";
+import { editFlag } from "../../actions/flagActions";
 
 const FIELD_DIV_CSS = "mt-2.5 flex justify-start items-center";
 const LABEL_CSS = "inline-block w-1/3 text-right mr-5";
